requirements.txt
pydantic
pydantic-settings
numpy
tiktoken
openai
pandas
datasets
together
nltk
google-genai
rouge-score
scipy
